After 82 - Interacting with Dialogs in MVVM - The Problem with Dialogs in the ViewModel

+++ FriendStorage/FriendStorage.UI/ViewModel/FriendEditViewModel.cs
@@ -6,6 +6,7 @@ using FriendStorage.UI.Wrappers;
 using Prism.Events;
 using System;
 using System.ComponentModel;
+using System.Windows;
 using System.Windows.Input;
 
 namespace FriendStorage.UI.ViewModel
@@ -55,8 +56,12 @@ namespace FriendStorage.UI.ViewModel
 
         private void OnDeleteExecute(object obj)
         {
-            _dataProvider.DeleteFriend(Friend.Id);
-            _eventAggregator.GetEvent<FriendDeletedEvent>().Publish(Friend.Id);
+            if (MessageBox.Show("Do you really want to delete the friend?",
+                "Delete friend", MessageBoxButton.YesNo) == MessageBoxResult.Yes)
+            {
+                _dataProvider.DeleteFriend(Friend.Id);
+                _eventAggregator.GetEvent<FriendDeletedEvent>().Publish(Friend.Id);
+            }
         }
 
         public void Load(int? friendId)
